+++ weave-js/src/components/PagePanelComponents/Home/Browse3/SelectOpVersion.tsx
@@ -5,8 +5,8 @@ import styled from 'styled-components';
 
 import {Pill} from '../../../Tag';
 import {Timestamp} from '../../../Timestamp';
-import {OpVersionSchema} from './pages/wfReactInterface/interface';
 import {opVersionKeyToRefUri} from './pages/wfReactInterface/utilities';
+import {OpVersionSchema} from './pages/wfReactInterface/wfDataModelHooksInterface';
 
 const Option = styled.li`
   display: flex;
