+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/wfReactInterface/utilities.ts
@@ -12,8 +12,8 @@ import {
   ObjectVersionSchema,
   OpCategory,
   OpVersionKey,
-} from './interface';
-import {CallSchema, Loadable} from './interface';
+} from './wfDataModelHooksInterface';
+import {CallSchema, Loadable} from './wfDataModelHooksInterface';
 
 type RefUri = string;
